requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
PyGithub>=2.3.0
gitpython>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
docker>=7.0.0
rich>=13.0.0
typer>=0.12.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
